Fiks posisjon av knapper i redigering av brev
